moved LoginForm into LoginPage
