+++ src/components/LoginPage/LoginPage.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import LoginForm from '../LoginForm/LoginForm';
+import LoginForm from './LoginForm';
 import { useHistory } from 'react-router-dom';
 
 // Material-UI
